Fix GetAudioEncoderConfigurations media ver20
